fixed UI:replaced import/export button icons to more fit the existing UI
